Expose method to get org.gnome.gdk.Window underlying a Widget
Add getWindow() to Widget, returning an org.gnome.gdk.Window. Add
considerable documentation to make it clear why this name is in use
and how to code it properly.

In order to get at the underlying GdkWindow, we need acceess to the
public window field in the C GtkWidget object which is of type
GdkWindow*. As we don't have generator code to do this at present, add
new class GtkWidgetOverride to provide the translation and JNI layers
to get at this attribute.