Expose maximize capability on Window
Add setMaximize() to Window, along with documentation of how to use
it and how it is differentiated from fullscreen mode.

Add getMaximize() to Window. The code to determine whether or not the
Window is maximized uses the WindowState flags object in GdkWindow, so
exposed in previous commits today. This technique was inspired by code
in devhelp's src/dh-window.c, line 703.