Add a method to allocate blank TreeIters
The various TreeModel, ListStore, etc methods that deal with rows and
columns need a TreeIter object; several methods return new ones, and
so we need a way to instantiate them.

C side, the usage is usually to declare a GtkTreeIter as a local
variable and then just pass it into the various GtkTreeModel methods
using the & operator, thus facilitting out-parameter behavior. This
doesn't work in the bindings context, however. So in an override
class, define a gtk_tree_iter_new() which declares a GtkTreeIter and
then uses gtk_tree_iter_copy() to allocate one and return a pointer to
it in a way compatible with the eventual TreeIter.release() which will
call gtk_tree_iter_free().

Expose this via TreeIter.<init>() but restrict it to package
visibility.

Allocating blank boxeds is probably something the code generator
should do for us.