Expose String methods on Value
Quite relucantly, expose <init>(String) and getString() on the Value
class. This is necessary because the setValue() and getValue() methods
in the TreeView family need to pass GValues in carrying the parameter.

Of course, if someone can invent a better way to do this that doesn't
involve making these methods public, please let us know.