Preliminary work on ListStore get and set values.
Start experimenting with ListStore, the simpler of the two TreeModels.
Concentrate work on the setValue() and getValue() cases, which we
expose in a strongly typed manner.

Implement a manual tranlation class, GtkTreeModelOverride, to get
around the problem of allocating a GValue necessary and then returning
it. Bit of a pain.

The ListStore constructor and column-to-type mapping is still
undefined, so implement a dummy constructor so we can at least
experiment with methods at runtime.

To that end, add a test case, ValidateTreeModel, to exercise the
TreeIter code as exposed by the getIterFirst() method on TreeModel. We
needed ListStore's append() too.