Compress TreeModel setValue()/getValue() signatures
We had started down the road of getValueString() and setValueString()
to avoid the problem of not being able to overload return types of
methods with the same signature (ie, our getters).

Using strongly typed DataColumnString, DataColumnReference, etc
instead of int (and instead of just DataColumn) means we can just keep
things simple and use the name getValue() and setValue().