Introduce new API connecting TreeView, TreeViewColumn, and CellRenderers
In the underlying library, TreeViewColumn has a method
addAttributeMapping() which is the mechanism used to tie a
CellRenderer to a column in a TreeModel. In the same way that GObject
properties are terribly from a Java perspective due to using String
identifiers rather than some strongly typed mechansim, the attributes
on CellRenderers are specified by Strings. Again, this is not
something we want to expose. After considerable exploration, however,
we might have found an API that would be clean, easy to use, type
safe, and not look cumbersome:

Expose CellRenderer attributes as setters on the CellRenderers
themselves. [They are, after all GObject properties internally].
Each setter in the CellRenderers takes a DataColumn as its argument
and then calls TreeViewColumn's addAttributeMapping() inside.

To make that work the CellRenderer constructor takes the
TreeViewColumn to which is going to be a part, stores a reference to
it, and then to save trouble automatically calls the TreeViewColumn
method to pack itself into it.

Likewise, to simplify the API we have the appendColumn() method on
TreeView instantiate and return a new TreeViewColumn in the process of
appending it.

The end result is that the same calls that one would have to make in
the underlying library are made, and with the same classes, but with a
somewhat simpler access path.

Finally, realizing that some attributes want to be set for all rows,
use the GObject property mechanism available from
org.gnome.glib.Object to expose setters on CellRenderers for those
attributes (eg, alignment) that should just be applied to all rows and
don't need to be driven by a DataColumn.