Fix bug in GValue.g_value_init(float)
Stupid cut and paste error in the JNI code implemeting our
g_value_init(float) call. Amazing that we didn't catch it before since
it's used in setAlignment(float,float) methods [looks like FALSE and 0
are the same thing, so we didn't notice], but that's what we get for
not having written proper unit tests for these code paths.

So, add test cases to ValidateProperties to check the [reluctantly]
public Value() constructors and their appropriate getters.