Expose setExpand() on TreeViewColumn and tune CellRenderer packing
Add the ability to set the expand property on TreeViewColumns. Having
taken some trouble to examine the behaviour of differnet setting
combinations, document the effect in particular on the last column.

Set the packing of CellRenderers to expand in their TreeViewColumn.
Since this is internal for us, it makes sense to ensure that the
CellRenderers get additional space - this also makes it so that right
aligned CellRenderers are actually on the right hand edge of their
vertical, rather than floating.