Example file demonstrating TreeView with a ListStore
Add ExampleTreeViewListStore in doc/examples/treeview/. This is a
simple but fully worked program that has been used to exercise and
test the TreeView and TreeModel APIs as we have been prototyping them.

In particular, this allowed us to evaluate the proposed design of
moving the attribute binding from TreeViewColumn to CellRenderer, and
allowed us to write the documentation for the sorting functionality we
have exposed.