Re-restrict visibility of Value methods
Add org.gnome.gtk.Value, package visible only, extending
org.gnome.glib.Value

It is somewhat cumbersome to have a wrapper class that does nothing
but chain up the constructors and getters, but this way the
constructors and getters in org.gnome.glib.Value can return to not
being publicly visible.

Thanks to Vreixo for helping work through the options.