Expose TreePath for use in TreeView's ROW_ACTIVATED signal
Add coverage for TreePath, an abstract way of representing a row in a
TreeModel. The constructor taking a String was stuck in missing.defs,
so move it to GtkTreePath.defs, and expose that constructor publicly,
along with an equals() override.

Add getIter(TreePath) to TreeModel, the method which converts between
abstract TreePaths and TreeIters tied to a specific model. Validate all
this with new test cases.

Expose signal ROW_ACTIVATED on TreeView. Test it at runtime by hooking
it up in ExampleTreeViewListStore.

Move the safety logic from the ListStore specific code in
GtkTreeModelOverride to the conversion functionin TreeModel, thereby
preventing NullPointerExceptions beating us to the punch when we mean
to throw an IllegalArgumentException. Replace it with some asserts
just to keep future development honest.