Initial coverage for TreeSelection
Begin implementing the code in TreeSelection, the helper class which
manages selection state in TreeViews. Expose getSelection() on
TreeView to get it.

Expose the constants in SelectionMode, used by TreeSelection's
setMode().

After reading the C source for GtkTreeSelection, add a (null-ok)
stanza to the .defs data for gtk_tree_selection_get_selected()'s
second parameter - it's a GtkTreeModel** which is just there for
convenience. We don't need it, and if it's NULL GtkTreeSelection just
ignores it. Means we can safe faffing about with the out parameter
arrays we would just be discarding anyway. Add tolerence for null
parameters to pointersOf() and objectsFor() in the Plumbing classes.