Preliminary Pixbuf coverage
Add initial constructors for Pixbuf, one taking a filename and the
other taking a filename and the scaling arguments. Had to bring them
across from the missing.defs file into GdkPixbuf.defs

Since the _new() functions use GError, we had to use a helper function
to catch the condition and rethrow as necessary. It's a bit messy.
Add a unit test checking the end result.

Expose setIcon() on Window taking a Pixbuf.