Fix references being stored in TreeModels
Switch to using a custom GBoxed of type BINDINGS_JAVA_REFERENCE.
Actually, this is _just_ a jobject pointer without a further structure
beneath, but registering this and using it as the GType for
java.lang.Objects being stored in TreeModels is sufficient for the
GValue system used by methods gtk_list_store_set_value() and
gtk_tree_model_get_value() to invoke our local
bindings_java_reference_copy() and bindings_java_reference_free()
functions and this is a simplistic but effective way of managing the
reference.

Thanks to Manish Singh for having suggested the approach of letting
the _copy() and _free() functions of a GBoxed manage the Java reference
to the stored object. He also indicated the precise pygobject code
which implements a comparable strategy for the Python bindings.

There was a terribly diffiuclt to debug data corruption associated
with trying to extract the stored reference. Ultimately we needed a
new pseudo helper function bindings_java_debug() to do a toString()
call for us [for some reason trying to do Class's getName() crashes
the VM]. Nevertheless, a JNI toString() equivalent was sufficient to
indicate that g_value_unset() invoking the DeleteGlobalRef() in
bindings_java_reference_free() was resulting in the jobject no longer
being usable. The solution, documented in GtkTreeModelOverride.c is to
create a local refence to the object before calling g_value_unset().

The helper function has been quietly stashed in
src/jni/bindings_java_util.c for now.