Add coverage for Format
Expose Format representing the cairo_format_t enum.

Adjust EnumBlock to change the way we compose the BindingsClass name.
We should probably apply this elsewhere.

Add an brutal kludge to TypeBlock to return the cairo package
statement instead of composing the package programatically as we do
for the G* libraries. This will need improving.