Move setBackground() to CellRenderer
Since CellRendererText "background" and CellRenderer's
"cell-background" do essentially the same thing, but the text one is
buggy (since it doesn't necessarily fill the entire height of the row)
and since the resultant mapping was setCellBackground() which was
stupid.

So refactor it so that setBackground() on CellRenderer sets the
"cell-background" property. Expose a setter for straight Strings as
well as the DataColumnString version.