Change design for TreeIter
The API to iterate through the rows in a TreeModel did not translate to Java
well at all. In order to fix this, we decided to move the iteration methods
[notably iterNext()] from TreeModel to TreeIterr. To make that work, we now
store a reference to the TreeModel in the TreeIter, while creating it.

Add a unit test to ValidateTreeModel to exercise the API.

Add a private getView() to TreeSelection in order to make it's getSelected()
[which returns a TreeIter] work.

This code done by Srichand Pendyala and Andrew Cowie pair programming in the
food court of some anonymous building in MIT, itself an anonymous place.