Expose VISIBILITY_NOTIFY_EVENT on Widget
Add coverage for the EventVisibility event wrapper and the
VisibilityState constants, as needed to expose VISIBILITY_NOTIFY_EVENT
on Widget.

This involved writing an override to set the events mask to generate
visibility events. This seems like a fairly unusual case, but if this
comes up again we'll have to generalize the mechanism.