Expose popup() on Menu
Add popup() to Menu to be used for hand-crafted popup context menus.
The underlying gtk_menu_popup() is insanely complicated with arguments
that are mostly redundant for the context popup case anyway. So
implement an override.

Add a second override for the case of popping up a context menu from a
StatusIcon, where there is a custom (*GtkMenuPositionFunc)() to be
used. Expose in the public API as popup(StatusIcon).