Expose modifyBackground() on Widget
Backport code adding coverage for the modifyBackground() method on
Widget from the 'treeview' branch. This necessitated adding coverage
of the Boxed type Color and the a series of constants in StateType.

Color as a boxed needs a creation method, but at this time the code
generator is not able to take care of this requirement. So implement
in an override (this pattern mirrors that used in TreeIter).