Expose VISIBILITY_NOTIFY_EVENT on Widget
Add coverage for the VISIBILITY_NOTIFY_EVENT signal on Widget. Add
.defs data for EventVisibility and write coverage for the
VisibilityState constants as used by this signal.

Note that the GDK flags for the GdkWindow have to be changed in order
for visibility events to be generated, so we do this in override code
before connecting the signal.

Backported from the 'treeview' branch.