Expose ResponseType to add type-safety to Dialog responses.
The ResponseType also presents common responses provided by Gtk+. This 
class is implemented in a different way that usual enums. This is needed 
for:

a) Deal with a enum that in the underlying Gtk+ code is treated as an 
int.
b) Allow users to define its own responses.

This commit also adds a method to add a Stock Button to a Dialog.