Expose MessageDialog and its methods.
Note that the usage of Pango markup has been exposed in a way similar to 
what we do in Label, instead of the original Gtk+ convenience methods 
that are variable length functions unsuitable for java.

This also adds the missing getPropertyBoolean to Object and the OTHER 
enum value in MessageType, also missing in the .defs data.