Refactor ResponseType code back to Constant hierarchy.
Refine code in ResponseType to revert to using the exising
infrastructure in the Constant (enum) hierarchy.

Alter GtkResponseType.defs to be a (define-flags ...) instead of a
(define-enum ...) to force the generated translation code to reach to
the native side for the ordinals, thus picking up the actual enum
values.

Expose the ordinal and instance infrastructure from Plumbing via a
GtkResponseTypeOverride extends Plumbing. This seems a number of extra
hops but the benefit is package visible methods on ResponseType that
are fairly clean when used by Dialog.

Create new unit test ValidateResponseType to verify the ordinal
handling and to make sure the custom subclassing works as advertised.