Introduce convenience subclasses for common MessageDialogs
Replace proposed static factory methods on MessageDialog with new
custom convenience subclasses InfoMessageDialog,
QuestionMessageDialog, ErrorMessageDialog, and WarningMessageDialog.
Doing this allows us to ensure the Dialogs are configured properly in
accordance with GNOME HIG standards and Desktop expectations.

As a bit of an experiment, add a few screenshots of what the different
configured custom MessageDialogs look like. This are manually
captured, of course.