Fix memory management bug relating to Window destruction
This is the fourth cut at this process. Based on advice from Owen
Taylor, revise the memory management of GtkWindows to leverage the
little known has_user_ref_count field to indicate to GTK that
something external is holding a ref count to the GtkWindow, rather
than itself.

Also on instantiation, hook up singal handlers for 'delete-event',
'hide', and 'show' to block deletion, unref, and ref respectively.