Add covarge for ImageMenuItem
Expose ImageMenuItem with the two constructors common to the other
MenuItem subclasses. Fix the GtkImageMenuItem.defs file's entry for
gtk_image_menu_item_new_from_stock() as NULL is a permissable value
for the second parameter. We're misusing it a bit at present, however

Add an ImageMenuItem to the ExampleSimpleMenu example.