Expose ability to set default icon for application's Windows
Expose gtk_window_set_default_icon() as Gtk.setDefaultIcon(). Yes, it
is notionally supposed to be a function on Window, but it has such
widespread impact that it seems better on Gtk. Furthermore, static
methods from Gtk.* are part of the API, whereas a single Window.blah()
is quite jarring compared to the rest of the Window methods which are
all instance methods.