Initial import gnome-screenshot code
This code from gnome-utils's gnome-screenshot/ directory, licenced
under GPLv2. These files are imported as-is from the upstream project
so we can tag this revision and branch from here to allow us to merge
future upstream changes. Next commits will morph the code into the
form we developed based on this work.