Initial import of automated screenshot mechanism
Introduce new top level class Harness to be what is run to present the
suite of Snap subclasses and capture a screenshot of each one.

Add base class Snapshot and first example class
SnapshotQuestionMessageDialog. There is still some confusion as to
what a good API will be for the Snap instances and I'm likewise not
terribly impressed with the machinery that is jammed here and there,
espectially in Snapshot.