Screenshot code licenced GPL only
The code from gnome-utils's gnome-screenshot is licenced under the
GPL. Therefore, use of org.gnome.screenshot.Screenshot implies that
the program as a whole must be licenced under a GPL compatible
licence (ie no Classpath Exception).

We've asked the upstream authors for a relicence, but until then,
annotate the Java classes and C code appropriately.