Add snapshots for Window and Button
Add SnapshotButton and SnapshotWindow for Button and Window
respectively, addingthem to those classes with <img> tags.

Add and reference a custom CSS stylesheet for the generated JavaDoc.
This is initially based on the default one from `javadoc` and modified
to have CSS for the img.snapshot class. Use class="snapshot" for the
added images.