Cleanup snapshot code
Add a snapshot for InfoMessageDialog, removing the fixed image that
was in previously in place in a doc-files/ directory.

Spell check new Java code and fix typos.

Add SnapshotDialog as intermediate class between Snapshot and
SnapshotInfoMessageDialog etc to provide for closing the window during
demo if the user clicks a button.