Add PixbufFormat constants
This is a workaround at present so we have a place to put strongly
typed constants representing the image formats we can actually write
to, thus allowing us to fix the signature of Pixbuf's save()

Add a short program in tests/exploration/ to run through the different
formats available and to determine which were writable. The lists
seemed pretty stable across several platforms, so we can go with
constants for the writable formats.

This will have to change to Boxed when we implement to code binding
GdkPixbufFormat properly.