Add a snapshot for TreeView
Add a Snapshot for use as an illustration in the documentation for
TreeView. Had to pick something as data, so I used the history of the
Canadian monarchy.

Expose setRulesHint() on TreeView.

Fix a bug in the screenshot Harness program - forgot to hide() the
previous snapshot. Oops.

Make a number of corrections an improvements to the JavaDoc for
DataColumnInteger, TreeView, and TreeModel while we're at it.