Corrections to FileChooser interface
As noted, and to achieve a GTK bugfix, remove internal
SELECTION_CHANGED from FileChooser and add FILE_SET to
FileChooserButton. This bumps the project dependency to GTK 2.12.

Expose setFilename(). It requires an absolute path, so expose the
parameter as type File. Add a test case to validate the setFilename()
behaviour.

Create a Snapshot to render a FileChooser example.