Expose setFilename() on FileChooser
Further attempts to get FileChooser right. While making a snapshot of
FileChooserDialog, add coverage for setFilename() on the FileChooser
interface.

Spent a lot of time messing around in unit tests relating to this. I
found it all a bit fragile, so we may have some problems lurking in
here. I nudged cycleMainLoop() and that seemed to help.

Decided to expose the signature as setFilename(String), not
setFilename(File); we still have to check for absolute pathnames
regardless and Strings are already in use in all the other methods.