Eliminate numerous warnings
Remove a now-obsolete tweak from the code generator which was
automatically importing the pubilc API class name to the generated
translation classes, several hundred of which (notably constants)
did not so require. [It would seem handling the reference to self in
parameter terms removed the need for importing the type]. This dropped
the warning count from several hundred to 65 or so.

Quietly redefine TreeModel as the object we are treating it as in the
.defs data. This stops the cast to Object extends Proxy from being
generated, and drops the warning count to 8.