Expose getToplevel() on Widget
Add coverage for Widget's getToplevel().

Add a unit test to verify its behaviour, as usual so we could document
it properly (and be aware of if the underlying library changes out
from under us).

Correct a JavaDoc linking bug and some typos. Add a blurb for
getParent() which is very much a related method.