Initial coverage of key values
Add class Keyval as a strongly typed wrapper around GDK's int keyvals,
allowing us to expose the business end of EventKey properly.

Supporting this we have corrected the .defs data for the gdk_keyval_*
functions and used the name() and fromName() functions to support
custom Keyvals and unit tests. Add test ValidateKeyboardHandling,
using a visibility workaround in GdkKeyvalOverride.

Keyvals are, for the moment, constructed with hardcoded values
extracted from the gtkkeysyms.h header file with a quick hack stashed
in src/util/keyval/.