Preliminary implementation of Plumbing for Cairo
First stab at the logic that will be required for the Cairo binding to
enable us to get arbitrary Proxies for a given pointer.

Add the beginnings of abstract type Pattern along with concrete
subclasses SolidPattern and SurfacePattern. Use this as the exercise for
implementing proxyFor(), native createPattern(), and implementing JNI
code.

After all this, I realized that I couldn't actually find an example of
a method that returned an abitrary Pattern to use in a unit test; the
only functions seemed to be constructors. So perhaps Pattern was a bad
example to start with, but I think the type-specific
create{Surface,Pattern,etc}() pattern set up in Plumbing.java along
with something that figures out the appropriate type to instantiate in
Plumbing.c is the right way to go.