Expose the basic capabilities necessary to manipulate Cursors
Add preliminary coverage of the Cursor and CursorType classes. The
constants in particular are only sparsly exposed but that's because
most of them are bogus anyway.

Add setCursor() on org.gnome.gdk's Window as that's how you manipulate
the Cursor appearing over a drawable.

For some reason the enum constants are missing from
src/defs/GdkCursorType.defs, so unfortunately I had to write them in
manually as needed.