Add coverage of TreeRowReference and supporting TreePath functionality
Expose TreeRowReference. Thankfully these have a proper _new()
function so we didn't have to mess around with the Boxed overrides
we've had to do elsewhere.

gtk_tree_row_reference_free() is most certainly not deprecated, so fix
that in the .defs data and call it from TreeRowReference.release().

Supporting this, add toString() in TreePath [which probably should
have been there long ago], clear() to ListStore, and getPath(TreeIter)
to TreeModel.

Write several more test fixtures into ValidateTreeModel to ensure that
TreeRowReference behaves as advertised and to make sure the various
supporting methods in TreePath work.