Expose ROW_CHANGED signal and fix bug in TreeIter
Add coverage for ROW_CHANGED on TreeModel.

We had originally removed the <init>(long) constructor from TreeIter,
thinking we had control over all the instantiation of the TreeIters
(necessary since we stash a reference to the source TreeModel
therein). However, it turns out that when handleRowActivated() is hit,
it is with a previously unseen TreeIter and boxedFor() will attempt to
create one. Blam.

Along the way, change the g_error() in Plumbing.c to g_critical().
NoSuchMethodError has already been thrown, and better to get a stack
trace, even though the program is in all liklihood borked at that
point [which, of course, is what g_critical() is for].