Fix usage of Value from Object wrapper
Bug fix misuse our GValue code from the setProperty() family of
methods in org.gnome.glib.Object. As we refactored things to handle
GValues of long (and thus dealt with the collision between long as
long and long as pointer), unit tests exposed a glitch whereby we had
missed a few cases and were using the wrong code path as a result.