Implement ComboBox and text specific subclass TextComboBox
Solidify API for our binding of ComboBox. Expose methods relating to
ComboBox being a CellLayout and something that works with TreeIters.

Extract and push down the text-only methods to new class TextComboBox 
in order to make ComboBox itself actually safe to use (otherwise, the
thing is bi-mode, with each mode being incompatible with the other.
Yuk!).

Unit tests for both ComboBox and TextComboBox.

Add a preliminary Snapshot, although this should probably be turned
into an example and a better image generated.