Implement ModifierType
Large patch making dramatic improvements to our binding of the
keyboard handling in GDK, including significant documentation
improvements across the key related classes.

Populate primary constants in class ModifierType, trying to make the
mappings as sensible as possible. Considerable work documenting it,
attempting to illustrate the sequences that occur when keys and
modifiers are pressed.

Take the code we used for testing and write it into new example
keyboard.ExampleSnooping in doc/examples/. This is still preliminary,
but at present it outputs the data that is described in the
ModifierType class JavaDoc comment.

Correct some of the constants in Keyval and KeypadKeyval.

Improve the discussion around Keyval's toUnicode(), warning of the
consequences of \0 chars. Add a tiny test fixture verifying these
statements.

Change the toString() override in Constant to only output the bare
class name, not the fully qualified one. It's an optics thing. Ammend
unit tests.