Add coverage for the Allocation class
Expose Allocation, the struct indicating the size and position of the
drawing area being given to a Widget as a result of the
request-allocation cycle.

Somewhat unusually, we have made a direct reference to the
GtkAllocation struct that lives in the GtkWidget struct. This should
have the benefit of meaning only one Allocation will be created (per
Widget) and that its getters will return correct current live values.
There are, obviously, some memory management risks but I think we've
daelt with that via a back-reference.