Fix corruption in snapshots with popups
There's something very screwy in the screenshot code; when there is a
popup it manages to corrupt the image data, even for PNGs that I
already thought were written. Clearly there is more going on than
meets the eye.

The workaround was to remove the popup() calls; this will need further
research. Perhaps manually downing the popups would do it.