Expose Statusbar
Introduce a new API for Statusbars, simplifying the hideous mess in
GTK down to a constructor and a single setMessage().

While documenting the use of setHasResizeGrip(), realized we needed
coverage of setResizable() in Window. Added.

While preparing the snapshot illustrating what a Statusbar looks like,
we needed a descriptive paragraph. Add coverage of setLineWrap() to
Label to facilitate this.