Implement a callback mechanism for TreeModelFilter
GtkTreeModelFilter has a function to hook up a callback function to
determine whether or not a row is visible. The problem is that the
argument to gtk_tree_model_filter_set_visible_func() is a function
pointer, (*GtkTreeModelVisibleFilterVisibleFunc), and we do not have
any mechanism to handle function pointers in the Java bindings.

We do, however, have a wizard mechanism for marshalling signal
callbacks. So, after considerable experimentation, when
setVisibleCallback() is called from Java, register a new custom signal
'visible' and then hook up a function which emits that signal when the
filter callback is invoked. Java side we can then implement that
"signal", returning boolean as appropriate.

Add the GInterface type to the various switches in our marshalling and
type conversion code as a second case alongside GObject.