Complete the .defs data for CairoStatus
Add the remaining constants to the .defs file for cairo_status_t. That
data is complete now.

Expose another constant in Status; I'm still not convinved we need
this class, or, at least, that asking people to check status all the
time is quite the right idea. I have a feeling that an unchecked
RuntimeExceptions should be thrown instead, although that would incur
an additional roundtrip per. Maybe calling Context's status() should
instead be a no-op vs throw Exception.