Add constructor for Drawables
Cook up a new constructor for Cairo Contexts taking a Drawable. This
is in turn a wrapper around gdk_cairo_create() which, after all,
returns a new cairo_t. So constructor it is. See the comments in the
CairoContextOverride files for details.

This also hoovers up a number of new methods related to Operator,
diagnostics like getStatus() returning the error state and utilities
like getTarget(). Some of these are still in flux, and documentation
is still incomplete.