New example of Cairo drawing in an EXPOSE_EVENT
Adjust the Cairo drawing example to do its work to a [GDK] Drawable
(ie, the source's GdkWindow) in an EXPOSE_EVENT handler hooked up to
that Widget.

Use an "empty" Image, adding a new constructor exposing the basic
gtk_image_new(). This gives one a blank slate to draw on when we call
source.getWindow() in the handler.

Move the original ExampleCairoDrawing to new TestCase
ValidateDrawingToFile, although that is not yet a complete or
functional test case.

Hook up the beginnings of dealing with the GdkRectangle in
EventExpose. This is another one of these structs within a struct, so
using the same trick as with Allocation and Requisition, return a
pointer to the data within the GdkEventExpose. This makes three times
we've had to do this by hand! And, note that getArea() in
GdkEventExpose is still blacklisted.