Temporarily disable use of Xvfb
For some reason the technique of using gnome-settings-daemon to
enact theme preferences in an Xvfb stopped working recently. Not sure
what the problem is, but in the mean time, hack it so that that part
of things is skipped, meaning that snapshots render to the developer's
screen. What a pain.

Touch up AboutDialog's snapshot.