Add source code examples
Flesh out the START page for the doc/examples/ section, and refer to
it from the /4.0/doc/ index page.

Assumes the source code files are rendered to HTML at
doc/example/package/ClassName.html, which is subject to change.
Additionally, generation of the HTML formatted versions will need to
be automated in the build system at some point.