Remove 15 warnings related to sentinels
We had a considerable number of compiler warnings about the fact that
our generated JNI code was making calls to GNOME functions taking
variable length argument lists without sufficient numbers of
arguments.

In most cases we aren't even calling these functions yet, but 
in almost all GNOME libraries, the convention is to terminate these lists with a
NULL parameter.

So we handle this by using the varargs marker in the defs data to add
an artificial last parameter (subsequently stripped off) which allows
us to know to add ", NULL" when generating the function call for that
FunctionBlock.