Preliminary attempt to do offscreen drawing
Although drawing in EXPOSE_EVENT handlers is what you are supposed to
do, it seems reasonable that we should be able to draw to a Pixmap
backend and then render that.

Code the very beginnings of Pixmap and an Image constructor to take
one.

There seem to be Colormap and Visual depth problems between creating
the Pixmap, creating a Cairo context on it, and then using that in the
Image, but the ExampleDrawingOffscreen is where you can exercise this.