Guess coverage extent
Add a simple scrpt to make first order approximation guess as to what
our coverage percentage is. This is very hackish, but gives a small
measure of progress. Note that we are not converging on the number of
GTK functions, but rather on the number of functions that need
exposing, which is an asymptote which is unfortunately an unknown
quantity.