Expose the ROW_EXANDED signal of TreeView.
This signal has the problem that it returns a TreeIter. Our generic signal 
handler has no way to initialize the model parameter, so we need to create a
custom handler to fill it properly before passing the iterator to the user.
Of course, this is handled internally and the user is not aware of this little
trick.