Expose RadioButton.
The implementation has some little deviations from Gtk+. First of them related
with a technical limitation of java-gnome infraestructure, as have no way to
wrap the GSList that represents the RadioButton group without writing an
override. It makes current RadioButton implementation unable to change the
group of a RadioButton, but given this feature seems to be quite useless,
I have chosen a simpler interface.

The other deviation is related to the introdution of new functions Gtk+ 
misses. One of them, getActiveButton() allows you to get the active Button in 
a group at any time, a feature really useful. The other is a signal that can 
be used to be notified when the active (selected) RadioButton in a group
changes, without needing to connect to the TOGGLE function of each button.
In my opinion both are very useful.