Further work to implement TreeModelSort.
Add setSortColumn() to the TreeSortable interface, then add the
implementation one-liner to TreeModelSort (the point of the exercise)
along with ListStore and TreeStore.

Implement SortType constants.

The fact that we needed all the error handling work in the previous
commit is because this is still not entirely documented correctly, and
problems can occur.