Remove direct GCJ build to native support
In response to an overwhelming number of requests for help whose root
cause turned out to be GCJ, remove it from the build system. It was
bit rotting anyway.

Some distros employ GCJ automatically and build a system wide
database of native files, so they will not be effected, and anyone can
always build the .jar file to native themselves if they have need.

Specifying the -C option to gcj makes it a .java to .class compiler,
and this is still supported via compiler=gcj, as is the GIJ
interpreter, via runtime=gij.

Be aware that using GCC's Java capabilities may still be a
non-starter; recent versions of GCJ still do not support
WeakReferences properly.

Thanks to everyone with Fedora systems (especially Red Hat staff at
the GTK hackfest in Berlin) who helped us work through this one.