Add support for i18n.
For message translation, we use a wrapper for C gettext(), thus keeping 
the usage of PO files and its "compilation" in .mo files. While GNU 
gettext has an alternative for Java language, we think it is better to 
keep things like they're usually done in GNOME environment.

We also incorporate in the translation function formatting capabilities, 
via the java MessageFormat.