Begin automating error handling
On advice from Carl Worth, compulsively check the cairo_status() after
each call into the library. Add call to checkStatus() in Context and
in Surface() and have it throw an exception if not SUCCESS. End result
is that status() no longer needs to be public on Context.

Further documentation improvements.

Typo and generic fixes.