Change type mapping cache in Plumbing
The original GType map in Plumbing was <String,Class>. This was
causing problems because the presence of a class [literal] does not
seem to be enough to force the class to actually be loaded.

The solution seems to be to use Class.forName(, true, ) rather
compulsively. This seems a bit inefficient; it sure would be nice if
a Class instance had a load() method. Presumably the internal
mechanism would be no worse than what is exposed as a result, but
still.

Vreixo Formoso suggested that ClassLoaders cache lookups. We'll see if
this becomes a problem area in profiling.

Also fix a few problems that were arising due to lacking .intern()
calls as were necessary given we are using an IdentityHashMap.