Add a test case that show a memory management problem.
The problem is related with GObjects created by Gtk+. When they are 
returned for first time, we create a Proxy for them, adding the toggle 
ref and removing the original ref. The problem is that in some functions 
we actually don't own that ref, so we are droping a Gtk+ ref. When the 
proxy goes out of scope, it has no ref from Gtk+ side (only the toggle 
ref), and it is freed. Next time we try to get the same object, an 
incorrect pointer is returned, thus leading to unexpected behavior, 
probably SEGFAULT.