Add list cleanup information to .defs data.
For methods that return a GList or GSList, we can have three situations.

1. Both the list and its contents are own by Gtk+, so we shouldn't free 
it.
2. The list is newly allocated, so we need to free it, but its items are 
own by Gtk+.
3. We own both the list and its contents.

However, our .defs data still not reflects that situation. In this 
commit we add caller-owns-return when needed. Note that we have added a 
new parameter for caller-own-return, #l, that means that only the list 
must be freed (2). For (3), we use the traditional #t parameter.