Fix bugs in gettext()
When there was no translation our JNI wrapper was returning garbage.
This appears to have been caused by the fact that gettext() returns
the supplied input string if no translation if found, but that we were
then calling ReleaseStringUTFChars() on. Surprising it didn't
segfault.

Add an escape checks on the Java side for the case of an empty String,
and in the C code, check to see if the char* returned by gettext() is
the input one - if so, return the original Java jstring directly.

Add unit tests verifying this behaviour.