Add ValidateInternationalization to UnitTests
Hardwire build/faster to generate the .mo file needed by the
ValidateInternationalization test case class.

Move the fr_CA.po dummy .po file to be alongside the .java file. These
two messages are not to be further translated! They are just there to
provide fixed results for the test cases. No need to have the PO
template in version control. Remvoed.

Modify setEnv() so that if null is passed as the value that variable
is removed from the environment via the unsetenv() function call. Add
unsetenv to JNI implementation.