Check that TreeIter passed to TreeModel functions is valid.
We have in TreeModel many functions that take a TreeIter. Of course, if 
the given TreeIter is not valid or belongs to another TreeModel, we will 
get a Gtk+ CRITICAL error, and a FatalError will be emitted. Obviously 
this is too severe error, so we try to catch possible problems and 
report them as an IllegalArgumentException.